♻️ refactor: 설정 - 계정 설정 기능 및 반복되는 스타일 분리
