--- apps/web/src/app/(needLogin)/mypage/setting/_components/SettingContainer/index.ts
@@ -1,2 +1,0 @@
-export * from './SettingContainer';
-export * from './SettingContainer.types';
