+++ apps/web/src/app/(needLogin)/mypage/setting/_components/index.ts
@@ -1,1 +1,1 @@
-export * from './SettingContainer';
+export * from './SettingContainer/SettingContainer';
