+++ apps/web/src/app/(needLogin)/mypage/setting/account/_components/ModifyForm/style.ts
@@ -1,0 +1,17 @@
+export const defaultStyle = {
+  width: '530px',
+  marginLeft: '0',
+  marginRight: '104px',
+  border: '1px solid #ccc',
+  borderRadius: '4px',
+  fontSize: '14px',
+};
+
+export const emailStyle = {
+  border: 'none',
+  background: '#eee',
+};
+
+export const nicknameConfirmStyle = {
+  marginRight: '16px',
+};
